chore: chart tooltips always using new tooltip
